dana-backend/requirements.txt
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
bs4==0.0.2
CacheControl==0.14.3
cachetools==5.5.2
catkin-pkg==1.0.0
cdo==1.6.1
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
coverage==7.9.2
cryptography==45.0.7
cycler==0.12.1
DateTime==5.5
distlib==0.3.9
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docutils==0.21.2
empy==4.2
firebase_admin==7.1.0
fonttools==4.59.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
lxml==5.3.2
matplotlib==3.10.5
msgpack==1.1.1
mysql-connector-python==9.2.0
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-repeat==0.9.4
pytest-rerunfailures==15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
redis==6.4.0
requests==2.32.3
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.31.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
threadpoolctl==3.6.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2025.2
ucimlrepo==0.0.7
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
